feat: add eslint and prettier tooling packages to monorepo (#279)
### TL;DR

This PR introduces a new `tooling` workspace along with various plugins and configurations for ESLint and Prettier.

Not used by other packages for now, that will come in later PRs

### What changed?

1. **Workspaces Update**: Added `tooling/*` to the list of workspaces in `package.json`.
2. **New tooling packages**:
   - **@isomer/eslint-config**: Added configurations and plugins for ESLint, including rules for TypeScript, React, and Next.js.
   - **@isomer/prettier-config**:  includes plugins for sorting imports and Tailwind CSS (if needed).
   - **@isomer/tsconfig**:  shared typescript configs for consistency.

### How to test?

1. Clone the repository.
2. Run `npm install` to install the new dependencies and set up the workspaces.

### Why make this change?

To streamline the development process by providing a consistent tsconfig, linting and formatting experience across the codebase. This will help in maintaining code quality and reducing the overhead of manually configuring these tools for each package.