feat(studio): use eslint, prettier, and tsconfig tooling packages in studio (#281)
### TL;DR

This PR replaces the current lint and formatter used in isomer studio with the eslint and prettier tooling packages, in order to maintain consistency of linting rules and formatting across the whole monorepo.

### What changed?

- Removed root `.prettierrc.json` file so prettier settings in repo is respected.
- Refactored existing codebase to conform with the new linting and formatting rules enforced by ESLint and Prettier.

### Why make this change?

The integration of Prettier with ESLint eliminates the need for a separate Prettier configuration file and ensures that code formatting rules are consistently applied across the entire codebase by using a single tool, ESLint, for both linting and formatting.