ISOM-1237: Allow edits to be saved as drafts (#271)
## Problem

During publishing flow, we need to support 2 modes of a resource: a main blob, a draft blob, and a state of if the resource is published or not.

We also only have an implicit way of knowing if a resource is a folder i.e. the absence of both mainBlob and draftBlob.

Closes ISOM-1237

## Solution

**Note this PR requires DB migration.**

Renames the current `blobId` field to be `mainBlobId` and introduces a new field called `draftBlobId`, both nullable columns.

It also adds a `isPublished` boolean to track if the resource is published. 

It adds a `isFolder` boolean to track if a resource is a folder. This should be a one time flag and hence we should not need to worry about frequent flag management.

We should return draft blobs preferentially to Studio to show the latest edits made by users. In the absence of a draft, we return the published version.

**Breaking Changes**

<!-- Does this PR contain any backward incompatible changes? If so, what are they and should there be special considerations for release? -->

- [X] Yes - this PR contains breaking changes
  - Details ...
- [ ] No - this PR is backwards compatible