feat: add new site side nav (#436)
This PR implements a new side navigation component (`CmsSideNav`) and integrates it into the CMS layout. The change updates the layout structure within the `CmsSidebarContainer`, adding support for a new sidebar section specifically for navigation purposes. Additionally, the PR includes the necessary modifications to the `AdminCmsSidebarLayout` to include the new side navigation.