feat: add collection resource list page and initial create collection page (#438)
This PR introduces several components and functionalities for managing collections in the dashboard. It includes the implementation of the `CollectionBanner`, `CollectionTable`, `CreateCollectionPageModal`, and various related components. Additionally, there are updates to the `Datatable` component to enhance functionality, and new utility hooks and schemas to support the creation and management of collection pages.

---