fix: crashes when there are other edits
previously, we crashed because we tried to read from undefined. this
resulted because we did a bad `setState` when we failed.

we should not set the whole page state because it is the whole blob.
since we send only the `content` (ie, the blocks) over to the backend,
we should set only that key.