fix: crashes when there are other edits (#650)
## Problem
previously, we crashed because we tried to read from undefined. this resulted because we did a bad `setState` when we failed.

we should not set the whole page state because it is the whole blob. since we send only the `content` (ie, the blocks) over to the backend, we should set only that key.

## Solution
1. set only the `content` key when we fail the mutation. this allows the toast to show and for the users to not lose data

## Screenshots

https://github.com/user-attachments/assets/77d3ce32-2c4e-4a9c-9380-7bf2d8876bf0