refactor(table): use dropdown list for table (#649)
* refactor: add table dropdown

adding the dropdown helps us to save space on the toolbar and makes it
more visually pleasing

* chore: remove toggle header row and col

we previously removed this because it led to bad interactions with
horizontal rules. a horizontal rule could split the table into 2 as it
has group: block but the table only allows text content.

this then means that the split table has a table cell (NOT header) on
the first row, which was previously not possible. this allowed us to add
a ordered list (allowed for table cell, not for table headers), then
toggle the cell to header, thus causing a crash

* chore: use vertical list for accordion