chore: update workflow for production (#672)
update so that we now deploy to production on release rather than using
a pr label. this is preferred because we don't have to make empty
commits and a fake pr to create a release.

removed the action to comment on pr because no easy out of the box
functionality was found. we can add it back next time if we feel that
it's needed