fix(permalinks): update sitemap generation for index pages (#723)
* refactor: page

1. update the getschema function so that we can use the `_index`
filepath. to do this, take note that the `_index` is only referenced at
build time or for the import of the json schema that is the actual index
page

* refactor: publishing

1. update index.ts so that generation of the sitemap trims any
terminating `_index`. this is allowed because we are going to update our
db so that folders have a `_index`.
2. for the types, we removed `undefined` for `fullPermalink` - this is
because the folders' index page and root page will have permalink:
_index

* refactor: update sitemap generation

1. filter out index pages for generating the localised sitemap. this is
because we already use it to get the id and we don't want to process it
again
2. filter out permalinks w `_index` when retrieving full permalink

* fix: shift error throwing into method

we cannot throw on the top level because we need to distinguish between
no results and having an array with a singular `_index` (as it gets
filtered out)

* feat: add new enum

1. added a new indexpage enum for resources
2. we need this because we want to differentiate the pages that are
index pages from those that are not

* chore: add check for index page type

* fix: add index page to types

* fix: resource

1. redirect index pages to `/page`
2. add index page to query

* chore: remove unused

* fix: add mapping for index page icon

* chore: update typings

* chore: format + lint

* fix: remove extra console log and fix docs