fix: build steps for turbo (#781)
## Problem
previously, we used to run the `template#build` command for builds of the template package. however, this is actually wrong. this is because the `build` command will be executed locally (and elsewhere) when we invoke `turbo build`. as part of the publishing flow, this package should only be built **after** the pages have been retrieved from teh database and written to disk.

because of this, we cannot actually use the `build` command to build this package. to preserve the correct dependency graph (and yet have this agnostic of `build`), i changed teh build command for templates to `build:template`

## Solution
1. update the `turbo.json` for `template` so that the task is `build:template` and update the `package.json` accordingly so that the `build` command is removed
2. update root `turbo.json` to add this as a command + also update the `package.json` so that turbo executes the `build:template` command